work in progress.

+++ frame/frame_rect.h
@@ -6,8 +6,8 @@
 /// @brief Struct which defines a rectangle in text mode.
 typedef struct
 {
-    uint8_t mTlx, mTly;
-    uint8_t mWidth, mHeight;
+    uint16_t mTlx, mTly;
+    uint16_t mWidth, mHeight;
 } frame_rect;
 
 #endif
